requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
numpy
tensorflow
pillow